upgrade to rest-js v2
