+++ package.json
@@ -25,7 +25,7 @@
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
     "eslint-plugin-react": "^7.3.0",
-    "jest": "^23.1.0",
+    "jest": "^23.6sw.0",
     "jest-cli": "^23.6.0",
     "rimraf": "^2.6.2",
     "webpack-dev-middleware": "^1.12.0",
@@ -36,6 +36,6 @@
     "babel-minify-webpack-plugin": "^0.3.0",
     "idempotent-babel-polyfill": "^0.1.1",
     "webpack": "^3.12.0",
-    "@esri/arcgis-rest-request": "^1.14.3"
+    "@esri/arcgis-rest-request": "^2.0.0"
   }
 }
