+++ src/helpers/__tests__/request-with-retry.exceptions.test.js
@@ -4,17 +4,14 @@ import {
   REQUEST_RETRY_CODES,
 } from '../../constants';
 
-import { request } from "@esri/arcgis-rest-request";
-
-jest.mock('request',
-  () => jest.fn().mockImplementation(() => ({ foo: 'bar' })),
-  { virtual: true },
-);
+jest.mock('@esri/arcgis-rest-request');
+import { request } from '@esri/arcgis-rest-request';
 
 describe('request with retry with exceptions', () => {
+  
   it('should be able to request with error', async () => {
     const responseData = { data: '1234' };
-    // const [request] = await esriLoader.loadModules(['esri/request']);
+
     request
       .mockImplementationOnce(() => {
         throw {
@@ -31,15 +28,14 @@ describe('request with retry with exceptions', () => {
     const r = await requestWithRetry(url, params, inputTime);
 
     expect(request).toHaveBeenCalledWith(url, {
-      foo: 'bar',
-      f: 'json',
-      responseType: 'json',
+      params: {
+        foo: 'bar'
+      }
     });
     expect(r).toEqual(responseData);
   });
 
   it('should not be able to request with other errors', async () => {
-    // const [request] = await esriLoader.loadModules(['esri/request']);
     request
       .mockImplementationOnce(() => {
         throw {
